+++ src/ansys/fluent/core/launcher/standalone_launcher.py
@@ -373,9 +373,10 @@ class StandaloneLauncher:
             logger.error(f"Exception caught - {type(ex).__name__}: {ex}")
             raise LaunchFluentError(self._launch_cmd) from ex
         finally:
-            server_info_file = Path(self._server_info_file_name)
-            if server_info_file.exists():
-                server_info_file.unlink()
+            if self.argvals.get("cleanup_on_exit", True):
+                server_info_file = Path(self._server_info_file_name)
+                if server_info_file.exists():
+                    server_info_file.unlink()
 
     @staticmethod
     def _get_journal_file_names(
